pyproject.toml
[project]
name = "openmemory"
version = "0.1.0"
description = "A persistent, semantic memory system for AI agents"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pydantic-settings[yaml]>=2.0",
    "httpx>=0.27",
    "numpy>=1.26",
    "sqlite-utils>=3.36",
    "python-dateutil>=2.9",
    "tiktoken>=0.7",
    "rich>=13.0",
    "PyYAML>=6.0",
]

[project.optional-dependencies]
local = [
    "sentence-transformers>=3.0",
]
langchain = [
    "langchain-core>=0.2",
]
mcp = [
    "mcp>=1.0",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
    "mypy>=1.10",
]

[project.scripts]
openmemory = "openmemory.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["openmemory"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.mypy]
python_version = "3.10"
strict = false
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]